requirements.txt
Babel==2.6.0
Django==2.0.8
django-filter==2.0.0
django-phonenumber-field==2.1.0
djangorestframework==3.9.0
Markdown==3.0.1
phonenumbers==8.10.2
Pillow==5.3.0
pytz==2017.3
